# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library API",
        "description": "API d'authentification et gestion de livres (Laravel + Sanctum).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local"
        }
    ],
    "paths": {
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Index (liste paginée)",
                "operationId": "5d5cb268f34aad5d87c4d6006448c123",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "id": 11,
                                            "title": "Clean Code",
                                            "author": "Robert C. Martin",
                                            "summary": "A handbook of agile software craftsmanship.",
                                            "isbn": "9780132350884",
                                            "_links": {
                                                "self": "http://localhost:8000/api/books/1",
                                                "update": "http://localhost:8000/api/books/1",
                                                "delete": "http://localhost:8000/api/books/1",
                                                "all": "http://localhost:8000/api/books"
                                            }
                                        }
                                    ],
                                    "links": {
                                        "first": "...",
                                        "last": "...",
                                        "prev": null,
                                        "next": "..."
                                    },
                                    "meta": {
                                        "current_page": 1,
                                        "per_page": 2,
                                        "total": 10
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Store (création)",
                "operationId": "7ab92284110ace5d0f7de52f372e70c0",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer 1|xxxxxxxxxxxxxxxxxxxxxxxx"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "title": "Clean Architecture",
                                "author": "Robert C. Martin",
                                "summary": "A guide to software structure and design.",
                                "isbn": "9780134494166",
                                "_links": {
                                    "self": "http://localhost:8000/api/books/1",
                                    "update": "http://localhost:8000/api/books/1",
                                    "delete": "http://localhost:8000/api/books/1",
                                    "all": "http://localhost:8000/api/books"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK (créé)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "data": {
                                        "id": 12,
                                        "title": "Clean Architecture",
                                        "author": "Robert C. Martin",
                                        "summary": "A guide to software structure and design.",
                                        "isbn": "9780134494166",
                                        "_links": {
                                            "self": "http://localhost:8000/api/books/1",
                                            "update": "http://localhost:8000/api/books/1",
                                            "delete": "http://localhost:8000/api/books/1",
                                            "all": "http://localhost:8000/api/books"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/books/{book}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Show (détail)",
                "operationId": "e2b7b266227828dd53fe29fa36552b23",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "book",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 12
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "data": {
                                        "id": 12,
                                        "title": "Clean Architecture",
                                        "author": "Robert C. Martin",
                                        "summary": "A guide to software structure and design.",
                                        "isbn": "9780134494166",
                                        "_links": {
                                            "self": "http://localhost:8000/api/books/1",
                                            "update": "http://localhost:8000/api/books/1",
                                            "delete": "http://localhost:8000/api/books/1",
                                            "all": "http://localhost:8000/api/books"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update",
                "operationId": "478701c722591c0ba42e99d37ec56ad7",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer 1|xxxxxxxxxxxxxxxxxxxxxxxx"
                    },
                    {
                        "name": "book",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 12
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "title": "Clean Architecture (2nd Edition)",
                                "author": "Robert C. Martin",
                                "summary": "Updated guide.",
                                "isbn": "9780134494166"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "data": {
                                        "id": 12,
                                        "title": "Clean Architecture (2nd Edition)",
                                        "author": "Robert C. Martin",
                                        "summary": "Updated guide.",
                                        "isbn": "9780134494166",
                                        "_links": {
                                            "self": "http://localhost:8000/api/books/1",
                                            "update": "http://localhost:8000/api/books/1",
                                            "delete": "http://localhost:8000/api/books/1",
                                            "all": "http://localhost:8000/api/books"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Destroy",
                "operationId": "0780eb2036bed1adf287d40708a9c033",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer 1|xxxxxxxxxxxxxxxxxxxxxxxx"
                    },
                    {
                        "name": "book",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 12
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "UnauthenticatedError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "NotFoundError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "No query results for model [App\\Models\\Book] 999."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Sanctum token. Example: Authorization: Bearer <token>",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}